pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pdf-manipulation"
version = "0.0.1"
description = "A python package for manipulating PDFs"
authors = [
    { name = "Karthik Vasu", email = "karthikvasu09@gmail.com" },
]
keywords = []
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
]
dependencies = [
    "PyPDF2 >= 3.0.1",
    "numpy >= 1.26.4",
    "opencv-python >= 4.10.0.82",
    "pillow >= 10.3.0",
]
requires-python = ">=3.7"

[project.urls]
"Source" = "https://github.com/kvasu2/PDF-manipulation"
